fix(console): address CodeRabbit review findings
- Increase step timeouts in tests-console steps (goto 30s needs 45s outer)
- Separate JSON.parse errors from handler exceptions in subscribeToSSE
- Drain stderr from cucumber subprocess to prevent backpressure
- Wrap coverage JSON.parse in try-catch against malformed file
- Close SSE subscription before nulling ref on terminal events
- Extract LiveRunView to its own .tsx file per one-component-per-file rule
