+++ e2e/tests/steps/tests-console.steps.ts
@@ -5,9 +5,9 @@ import { config } from '../support/types.js'
 
 When(
   'I open the tests page',
-  { timeout: 20_000 },
+  { timeout: 45_000 },
   async function (this: AgentWorld) {
-    await this.page.goto(`${config.consoleUrl}/tests`)
+    await this.page.goto(`${config.consoleUrl}/tests`, { timeout: 30_000 })
     await this.page
       .getByRole('button', { name: 'Run tests' })
       .waitFor({ state: 'visible', timeout: 15_000 })
@@ -16,10 +16,10 @@ When(
 
 Then(
   'the {string} button becomes enabled',
-  { timeout: 30_000 },
+  { timeout: 90_000 },
   async function (this: AgentWorld, buttonName: string) {
     const button = this.page.getByRole('button', { name: buttonName })
-    await expect(button).toBeVisible({ timeout: 30_000 })
-    await expect(button).toBeEnabled({ timeout: 30_000 })
+    await expect(button).toBeVisible({ timeout: 60_000 })
+    await expect(button).toBeEnabled({ timeout: 60_000 })
   }
 )
